server/requirements.txt
Flask
python-docx
docx2pdf
PyYAML
